javascript for cart

+++ shopping_site/shop/models.py
@@ -6,7 +6,7 @@ class Product(models.Model):
     product_name = models.CharField(max_length=150)
     category = models.CharField(max_length=100, default='')
     sub_category = models.CharField(max_length=100, default='')
-    price = models.FloatField(default=0)
+    price = models.DecimalField(max_digits=100, decimal_places=2 ,default=0)
     desc = models.CharField(max_length=200, default='', null=True, blank=True)
     date = models.DateTimeField(auto_now_add=True, auto_now=False)
     updated = models.DateTimeField(auto_now_add=False, auto_now=True)
